Creating abstract parent for locked presenters
