+++ app/src/main/java/mozilla/lockbox/presenter/AppLockedPresenter.kt
@@ -9,33 +9,43 @@ package mozilla.lockbox.presenter
 import io.reactivex.Observable
 import io.reactivex.rxkotlin.addTo
 import mozilla.lockbox.R
-import mozilla.lockbox.action.AutofillAction
 import mozilla.lockbox.action.DataStoreAction
 import mozilla.lockbox.action.FingerprintAuthAction
 import mozilla.lockbox.action.RouteAction
+import mozilla.lockbox.action.UnlockingAction
 import mozilla.lockbox.flux.Dispatcher
-import mozilla.lockbox.flux.Presenter
 import mozilla.lockbox.store.FingerprintStore
 import mozilla.lockbox.store.LockedStore
 import mozilla.lockbox.store.SettingStore
+import mozilla.lockbox.support.isTesting
 import java.util.concurrent.TimeUnit
 
-interface AutofillLockedView {
-    fun unlockFallback()
+interface LockedView {
+    val unlockButtonTaps: Observable<Unit>
     val unlockConfirmed: Observable<Boolean>
 }
 
-class AutofillLockedPresenter(
-    private val lockedView: AutofillLockedView,
+class AppLockedPresenter(
+    private val view: LockedView,
     private val dispatcher: Dispatcher = Dispatcher.shared,
     private val fingerprintStore: FingerprintStore = FingerprintStore.shared,
-    private val settingStore: SettingStore = SettingStore.shared,
-    private val lockedStore: LockedStore = LockedStore.shared
-) : LockedPresenter() {
+    private val lockedStore: LockedStore = LockedStore.shared,
+    private val settingStore: SettingStore = SettingStore.shared
+) : LockedPresenter(dispatcher, fingerprintStore) {
+
+    private val delay: Long = if (isTesting()) 0 else 1
+
     override fun onViewReady() {
-        settingStore.unlockWithFingerprint
-            .take(1)
-            .delay(500, TimeUnit.MILLISECONDS)
+        Observable.just(Unit)
+            .delay(delay, TimeUnit.SECONDS)
+            .switchMap { lockedStore.canLaunchAuthenticationOnForeground.take(1) }
+            .filter { it }
+            .map { Unit }
+            .mergeWith(view.unlockButtonTaps)
+            .doOnNext {
+                dispatcher.dispatch(UnlockingAction(true))
+            }
+            .switchMap { settingStore.unlockWithFingerprint.take(1) }
             .subscribe {
                 if (fingerprintStore.isFingerprintAuthAvailable && it) {
                     dispatcher.dispatch(RouteAction.DialogFragment.FingerprintDialog(R.string.fingerprint_dialog_title))
@@ -45,29 +55,25 @@ class AutofillLockedPresenter(
             }
             .addTo(compositeDisposable)
 
+        view.unlockConfirmed
+            .filter { it }
+            .subscribe {
+                unlock()
+            }
+            .addTo(compositeDisposable)
+
         lockedStore.onAuthentication
             .subscribe {
                 when (it) {
                     is FingerprintAuthAction.OnSuccess -> unlock()
                     is FingerprintAuthAction.OnError -> unlockFallback()
-                    is FingerprintAuthAction.OnCancel -> dispatcher.dispatch(AutofillAction.Cancel)
                 }
             }
-            .addTo(compositeDisposable)
-
-        lockedView.unlockConfirmed
-            .map {
-                if (it) {
-                    DataStoreAction.Unlock
-                } else {
-                    AutofillAction.Cancel
-                }
-            }
-            .subscribe(dispatcher::dispatch)
             .addTo(compositeDisposable)
     }
 
     override fun unlock() {
         dispatcher.dispatch(DataStoreAction.Unlock)
+        dispatcher.dispatch(UnlockingAction(false))
     }
 }
